lots of css

+++ client/src/actions/admin.js
@@ -38,6 +38,28 @@ export const addTutorial = formData => async dispatch => {
     }
 };
 
+// DELTE specific video
+export const deleteVideo = (id, vidId) => async dispatch => {
+    try {
+        const res = await axios.delete(`/api/tutorial/videos/${id}/${vidId} `);
+
+        dispatch({
+            type: DELETE_VIDEO,
+            payload: res.data
+        });
+
+        dispatch(setAlert('Tutorial Removed', 'success'));
+    } catch (err) {
+        dispatch({
+            type: TUTORIAL_ERROR,
+            payload: {
+                msg: err.response.statusText,
+                status: err.response.status
+            }
+        });
+    }
+};
+
 // Get all tutorials
 export const getTutorials = () => async dispatch => {
     try {
@@ -61,26 +83,26 @@ export const getTutorials = () => async dispatch => {
 
 // DELETE SINGLE TUTORIAL
 export const deleteTutorial = id => async dispatch => {
-    // if (window.confirm('Are you sure? This CANNOT be undone!')) {
-    try {
-        const res = await axios.delete(`/api/tutorial/${id} `);
+    if (window.confirm('Are you sure? This CANNOT be undone!')) {
+        try {
+            const res = await axios.delete(`/api/tutorial/${id} `);
 
-        dispatch({
-            type: DELETE_TUTORIAL,
-            payload: res.data
-        });
+            dispatch({
+                type: DELETE_TUTORIAL,
+                payload: res.data
+            });
 
-        dispatch(setAlert('Tutorial Removed', 'success'));
-    } catch (err) {
-        dispatch({
-            type: TUTORIAL_ERROR,
-            payload: {
-                msg: err.response.statusText,
-                status: err.response.status
-            }
-        });
+            dispatch(setAlert('Tutorial Removed', 'success'));
+        } catch (err) {
+            dispatch({
+                type: TUTORIAL_ERROR,
+                payload: {
+                    msg: err.response.statusText,
+                    status: err.response.status
+                }
+            });
+        }
     }
-    // }
 };
 
 // Get single tutorial
